+++ spring-cloud-tencent-examples/quickstart-example/quickstart-caller-service/src/main/java/com/tencent/cloud/quickstart/caller/CustomMetadataProvider.java
@@ -18,6 +18,9 @@
 
 package com.tencent.cloud.quickstart.caller;
 
+import java.util.HashMap;
+import java.util.Map;
+
 import com.tencent.cloud.common.spi.InstanceMetadataProvider;
 
 import org.springframework.stereotype.Component;
@@ -28,6 +31,12 @@ import org.springframework.stereotype.Component;
 @Component
 public class CustomMetadataProvider implements InstanceMetadataProvider {
 
+	@Override
+	public Map<String, String> getMetadata() {
+		Map<String, String> metadata = new HashMap<>();
+		metadata.put("k1", "v1");
+		return metadata;
+	}
 //	@Override
 //	public String getRegion() {
 //		return "huadong";